Move 32-bit float into primitives
