+++ library/Rattletrap/AttributeValue/WeldedInfo.hs
@@ -1,6 +1,6 @@
 module Rattletrap.AttributeValue.WeldedInfo where
 
-import Rattletrap.Float32
+import Rattletrap.Primitive.Float32
 import Rattletrap.Int32
 import Rattletrap.Int8Vector
 import Rattletrap.Vector
